Refactored the format string parameter detection into its own file in the utils folder. This way additional checks can benefit from its functionality.
