+++ src/cwe_checker_lib/src/utils/mod.rs
@@ -1,5 +1,6 @@
 //! This module contains various utility modules and helper functions.
 
+pub mod arguments;
 pub mod binary;
 pub mod graph_utils;
 pub mod log;
